examples/database-app/src/web/requirements.txt
Flask==3.0.0
pyodbc==5.0.1
opencensus-ext-azure==1.1.13
gunicorn==21.2.0
